[ADD] Created Password component
